requirements.txt
python-dotenv~=1.0.0
httpx~=0.25.2
aiogram~=3.2.0